Re-add business-form-options file, update day and month inputs to form-control-select with prefilled options
