+++ packages/webcomponents/src/components/business-forms/business-form/business-core-info/business-core-info.tsx
@@ -1,5 +1,5 @@
 import { Component, Host, h, Prop, State } from '@stencil/core';
-import { BusinessTypeOptions } from '../../utils/business-form-types';
+import { BusinessTypeOptions } from '../../utils/business-form-options';
 import { FormController } from '../../../form/form';
 import { PHONE_MASKS, TAX_ID_MASKS } from '../../../../utils/form-input-masks';
 import { CoreBusinessInfo, ICoreBusinessInfo } from '../../../../api/Business';
